+++ example/lib/theme_mode_builder_example_app.dart
@@ -48,77 +48,84 @@ class _HomeState extends State<Home> {
           'Theme Mode Builder Example',
         ),
       ),
-      body: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: <Widget>[
-          /// text to show current mode
-          Text(
-            "Current Mode: ${ThemeModeBuilderConfig.getThemeMode().toString().split(".").last.toUpperCase()}",
-            style: const TextStyle(
-              fontSize: 25,
-              fontWeight: FontWeight.bold,
-            ),
-          ),
-          const SizedBox(height: 20),
-
-          /// toggle functionality
-          ElevatedButton(
-            onPressed: () async {
-              await ThemeModeBuilderConfig.toggleTheme();
-            },
-            child: const Text(
-              'Toggle Theme',
-            ),
-          ),
-
-          /// spacing to make things look cool!
-          const SizedBox(height: 15),
-
-          /// separate buttons functionality
-          Row(
+      body: SafeArea(
+        child: Padding(
+          padding: const EdgeInsets.all(16),
+          child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
+            crossAxisAlignment: CrossAxisAlignment.stretch,
             children: <Widget>[
-              /// changes theme to light mode
-              ElevatedButton.icon(
-                icon: const Icon(Icons.phone_android),
-                label: const Text(
-                  "System Mode",
+              /// text to show current mode
+              Text(
+                "Current Mode: ${ThemeModeBuilderConfig.getThemeMode().toString().split(".").last.toUpperCase()}",
+                style: const TextStyle(
+                  fontSize: 25,
+                  fontWeight: FontWeight.bold,
                 ),
+                textAlign: TextAlign.center,
+              ),
+              const SizedBox(height: 20),
+
+              /// toggle functionality
+              ElevatedButton(
                 onPressed: () async {
-                  await ThemeModeBuilderConfig.setSystem();
+                  await ThemeModeBuilderConfig.toggleTheme();
                 },
+                child: const Text(
+                  'Toggle Theme',
+                ),
               ),
 
               /// spacing to make things look cool!
-              const SizedBox(width: 15),
+              const SizedBox(height: 15),
 
-              /// changes theme to darl mode
-              ElevatedButton.icon(
-                icon: const Icon(Icons.lightbulb),
-                label: const Text(
-                  "Light Mode",
-                ),
-                onPressed: () async {
-                  await ThemeModeBuilderConfig.setLight();
-                },
-              ),
+              /// separate buttons functionality
+              Wrap(
+                alignment: WrapAlignment.center,
+                children: <Widget>[
+                  /// changes theme to light mode
+                  ElevatedButton.icon(
+                    icon: const Icon(Icons.phone_android),
+                    label: const Text(
+                      "System Mode",
+                    ),
+                    onPressed: () async {
+                      await ThemeModeBuilderConfig.setSystem();
+                    },
+                  ),
 
-              /// spacing to make things look cool!
-              const SizedBox(width: 15),
+                  /// spacing to make things look cool!
+                  const SizedBox(width: 15),
 
-              /// changes theme to darl mode
-              ElevatedButton.icon(
-                icon: const Icon(Icons.lightbulb_outline),
-                label: const Text(
-                  "Dark Mode",
-                ),
-                onPressed: () async {
-                  await ThemeModeBuilderConfig.setDark();
-                },
+                  /// changes theme to darl mode
+                  ElevatedButton.icon(
+                    icon: const Icon(Icons.lightbulb),
+                    label: const Text(
+                      "Light Mode",
+                    ),
+                    onPressed: () async {
+                      await ThemeModeBuilderConfig.setLight();
+                    },
+                  ),
+
+                  /// spacing to make things look cool!
+                  const SizedBox(width: 15),
+
+                  /// changes theme to darl mode
+                  ElevatedButton.icon(
+                    icon: const Icon(Icons.lightbulb_outline),
+                    label: const Text(
+                      "Dark Mode",
+                    ),
+                    onPressed: () async {
+                      await ThemeModeBuilderConfig.setDark();
+                    },
+                  ),
+                ],
               ),
             ],
           ),
-        ],
+        ),
       ),
     );
   }