+++ lib/theme_mode_builder/theme_mode_builder.dart
@@ -1,16 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:hive_flutter/hive_flutter.dart';
 import 'package:provider/provider.dart';
 import 'package:theme_mode_builder/common/theme_mode_builder_config.dart';
-
-import 'theme_mode_builder_model.dart';
+import 'package:theme_mode_builder/services/theme_storage_service.dart';
 
 /// [ThemeModeBuilder] constructor, you need to provide the [builder].
 class ThemeModeBuilder extends StatelessWidget {
   /// [builder] requires you to return a [Widget] and provides two arguments:
   ///
   /// The two arguments are:
-  /// * [BuildContext], context; the thing that no body knows what it is
+  /// * [BuildContext], context; the thing that nobody knows what it is
   /// * [ThemeMode], tells you what is the current theme mode
   final Widget Function(BuildContext, ThemeMode) builder;
 
@@ -22,33 +20,23 @@ class ThemeModeBuilder extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return ChangeNotifierProvider<ThemeModeBuilderModel>(
-      create: (BuildContext context) {
-        final ThemeModeBuilderModel themeModeBuilderModel =
-            ThemeModeBuilderModel();
-
-        themeModeBuilderModel.init();
-
-        return themeModeBuilderModel;
-      },
-      builder: (
-        BuildContext context,
-        Widget? child,
-      ) {
-        return ValueListenableBuilder<Box<bool>>(
-          valueListenable:
-              context.read<ThemeModeBuilderModel>().themeBox!.listenable(),
-          builder: (
-            BuildContext? context,
-            Box<bool>? box,
-            Widget? child,
-          ) {
-            return builder(
-              context!,
-              ThemeModeBuilderConfig.getThemeMode(),
-            );
-          },
-        );
+    return FutureBuilder<void>(
+      future: ThemeModeBuilderConfig.ensureInitialized(),
+      builder: (BuildContext context, AsyncSnapshot<void> snapshot) {
+        if (snapshot.connectionState == ConnectionState.done) {
+          return ChangeNotifierProvider<ThemeStorageService>.value(
+            value: ThemeModeBuilderConfig.themeStorageService,
+            child: Consumer<ThemeStorageService>(
+              builder: (BuildContext context,
+                  ThemeStorageService themeStorageService, Widget? child) {
+                return builder(context, themeStorageService.getThemeMode());
+              },
+            ),
+          );
+        } else {
+          // You can return a loading widget while the initialization is in progress
+          return const Center(child: CircularProgressIndicator());
+        }
       },
     );
   }